move prompt input to its own pkg
