+++ pkg/util/prompt/prompt.go
@@ -1,0 +1,28 @@
+package prompt
+
+import (
+	"bufio"
+	"os"
+	"strings"
+)
+
+var inputScanner = bufio.NewScanner(os.Stdin)
+
+// SetInputScanner allows you to override the default input scanner with your own.
+func SetInputScanner(scanner *bufio.Scanner) {
+	inputScanner = scanner
+}
+
+// GetInput reads input from an input buffer and returns the result as a string.
+func GetInput(defaultValue string) string {
+	inputScanner.Scan()
+	input := inputScanner.Text()
+
+	// If the value from the input buffer is blank, then use the default instead.
+	value := strings.TrimSpace(input)
+	if value == "" {
+		value = defaultValue
+	}
+
+	return value
+}
